+++ lib/providers/song_provider.dart
@@ -2,13 +2,14 @@ import 'dart:io';
 
 import 'package:audio_session/audio_session.dart';
 import 'package:christian_lyrics/christian_lyrics.dart';
-import 'package:flow_lyrix/widgets/christian_lyrics_widgets.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:just_audio/just_audio.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:sylt_parser/sylt_parser.dart';
 import 'package:uri_to_file/uri_to_file.dart';
+
+import '../models/position_data.dart';
 
 class SongProvider {
   SongProvider() {